feat: AND-OR-Invert (AOI)
You are given a design board with three input pins A, B and C, and
one output pin R.
Build a circuit which raises R whenever B or C (or both) are up, but
A is down. Otherwise the circuit drives R down. In other words, R is 1
when B or C or both are 1 at the same time as A is 0, otherwise R is 0.
You may need more than one gate to achieve this.