feat:  3-input Majority Gate
You are given a design board with three input pins A, B and C, and
one output pin R.
Build a gate circuit which raises R whenever the majority of A, B and
C are up, and which keeps it down otherwise. In other words, R is 1
whenever at least two of A, B and C are 1, otherwise R is 0. This can
be done by following your intuition or if you prefer using the Minterm
Method