feat: sign extension circuit
You are given a design board with a 4-bit input pin A, and a 6-bit output pin R. The data on pin A are assumed to be in 4-bit 2’s complement
representation. Build a circuit that asserts the value represented by A
on pin R assuming 6-bit 2’s complement representation for R. In other
words, extend the input representation from 4 to 6 bit 2’s complement.