feat: 2-to-4 multiplexer
You are given a design board with two input pins, a 2-bit pin A and
a 1-bit pin enable and 4 output pins R0–R3. Build a gate circuit that
behaves as follows. If enable is high, the value on pin A selects which one
of the pins R0–R3 is raised. If enable is low, then all output pins should
be low.