feat: 4-to-16 decoder
You are given a design board with two input pins, a 4-bit pin A and a 1-
bit pin enable and 16 output pins R0–R15. Build a 4-to-16 decoder using
the 2-to-4 decoder you built earlier. (For your convenience we supplied
our own version with the design board but you are free to ignore it.)
Hint: use one 2-to-4 decoder to decode two most significant bits of A
into 4 enable signals, and use four further 2-to-4 decoders to decode the
two least significant bits in parallel taking those enable signals one each.