feat: 6-bit signed-number subtractor circuit
You are given a design board with two input pins A and B and three
output pins R, C.out and V. Build from gates a 6-bit signed-number
subtractor circuit, which subtracts B from A asserting the difference on
R and the carry-out on C.out. If subtraction causes a signed-number
overflow, your circuit must assert 1 on V, otherwise V should be low.
You may use — and combine — any of the circuits you have already
built as the basis for this, but you may not use any of Logisim’s builtin arithmetic devices to solve this problem.