feat: 6-bit ripple carry adder circuit
You are given a design board with three input pins A,B and C.in and
two output pins R and C.out. Build from gates a 6-bit adder, which adds
together A, B and carry-in C.in asserting the sum on R and the carry-out
on C.out.
You are given a 1-bit full adder to start you off. You may use this
circuit in any way you like, but you may not use any components from
Logisim’s arithmetic library Arithmetic.