feat: 6-bit input and output
You are given a design board with one 6-bit input pin A and one 6-bit
output pin R.
Build a circuit that behaves as follows. If bits 2 and 3 of A are equal,
assert the value on R which is the same as the value A except bit 5 is
negated. Otherwise assert the value of A on R as is.