feat: 4-Bit Reversible counter
You are given a design board with four input pins a 4-bit INDATA,
1-bit Load,Enable, and Clock; and one output, a 4-bit OUTDATA.
In this exercise you are building a reversible counter, i.e. a counter
that counts up to an upper limit, then switches to counting down and
carries on until it reaches a lower limit, at which point it switches back
to counting up, etc. Specifically the requirement is for a 4-bit reversible
counter which goes up to 1111 and down to 0. For example if we set
it to 5 initially (we use decimal numbers for ease of reading) then in
subsequent clock cycles, assuming the counting is enabled, the output
must be
6, 7, 8, . . . , 13, 14, 15, 14, 13, . . . , 2, 1, 0, 1, 2, . . .
Build a sequential circuit that contains a register (Don’t forget to
trigger that register by the FALLING edge of the clock, Logisim’s default
is the opposite!) for the current count C (4 bits) and another one (1 but)
for the direction of counting d (0: up, 1: down). Notice that it is very
easy to make 0001 and 1111 (−1 in 4-bit 2’s complement) from d using
a splitter. You will require an adder (use a library one).
It makes sense to connect the output of the d register back to its input
via a NOT gate, this creating a toggle. The toggle can be activated by
the register’s enable signal.
The main challenge left for you is to wire the enable network to the
toggle, which should trigger it whenever the count direction is up and
C = 14 and when the count direction is dowm and C = 1.
The other inputs have the following effect.
The register is updated every clock cycle in which Enable is up. If Load
is down, the counter is updated (in the appropriate direction), otherwise
it is loaded with the data asserted on the INDATA pin.
The register data output should be connected with the output pin
OUTDATA.