feat: custom design tailored
You are given a design board with two 1-bit input pins, Reset and
Clock; and one output, a 11-bit garland.
Build a driver for a 11-light garland for a Xmas tree. You will need a
11-bit register in your solution, whose output should be connected to pin
garland. ATTENTION! Logisim’s default for triggering is rising edge;
you need your registers to be triggered by the falling edge of the clock.
If the pin Reset is up in a clock cycle, your cicuit should latch the value
0x020 (a middle 1) in the register. If Reset is down, the register should be
asserted with a shift of the value towards the more senior bits, i.e. shift
left. If the shift results in shifting a ‘1’ out of the register, the 1 has to be
re-inserted at the junior end (bit 0).