feat: Xmas bouncing light - 2
You are given a design board with two 1-bit input pins, Reset and
Clock; and one output, a 11-bit garland.
In this exercise you are building a BOUNCING-light garland for a
Xmas tree.
You will need a 11-bit register in your solution, whose output should be
connected to pin garland. ATTENTION! Logisim’s default for triggering
is rising edge; you need your registers to be triggered by the falling
edge of the clock.
If the pin Reset is up in a clock cycle, your cicuit should latch the value
0x020 (a middle 1) in the register. If Reset is down, the register should
be asserted with a shift of the value towards the more senior bits, i.e.
shift left. If the shift would result in shifting a ‘1’ out of the register,
the direction of shift should change instead.
If Reset rises again in any clock cycle, this does not change the
current direction of shift, but it will move the ’1’ to the middle position
again.