feat: memory comparator
You are given a design board with three input pins a 6-bit Target, 1-bit
Load and Clock; and three outputs, an 8-bit Location and 1-bit Stop and
Found. You are also given a 32-cell ROM with some 6-bit content.
This exercise is about searching for data in Read-Only Memory (ROM).
Build a circuit that operates in two modes:
Load=1 The circuit latches the 6-bit value Target in a register on the
falling edge of the clock. We will call this register the target reg-
ister. The circuit also latches 1 in an 8-bit register intended for
addressing the ROM, which we will call the address register.
Load=0 As long as Load=0 the circuit examines the ROM cell whose
address is contained in the address register to determine if the cell’s
content is the same as the target register’s. If so, the Found pin is
raised, otherwise the address register is incremented. If the address
register contains 0xFF in any clock cycle, the pin Stop should be
raised in that cycle.