feat: Fibonacci Series
Build a circuit that computes all the Fibonacci numbers not exceeding
2 − 1. The Fibonacci numbers are a sequence where the first two num-
bers are 1,1, and every next number is the sum of its two predecessors.
Here is the beginning of the sequence:
32
1, 1, 2, 3, 5, 8, 13 . . .
Indeed 1 + 1 = 2, 1 + 2 = 3, 2 + 3 = 5, etc.
Mathematically speaking, you are producing the sequence
a1 , a2 , . . . ,
where a1 = 1,a2 = 1, and for all k > 2, ak = ak−1 + ak−2
Use a pair of 32-bit registers to hold the last two members of the
squence. You may utilise the Arithmetic library of Logisim for the 32-
bit adder (or anything else you want to use).
Your circuit has one input pin: the clock, and three output pins, a
6-bit k, a 32-bit a, and a single digit stop. The circuit should assert the
values k and ak , where k = 1, 2, . . . on the corresponding output pins
at the end of each consecutive clock cycle until the current Fibbonaci
number exceeds 232 − 1, at which point the stop signal must go up by
the end of the clock cycle.