feat: GCD
Your task is to build a circuit that computes the Greatest Common
Divisor (GCD) of two 8-bit unsigned integers, which are guaranteed to
be greater than zero. The solution in hardware follows the solution in
software you may have encountered earlier as it utilises Euclid’s algo-
rithm.
Euclid’s algorithm For two positive numbers A and B, repeat:
• if A=B, then the QCD=A, stop
• if A>B set A:=A−B
• if A<B set B:=B−A
Your task
Build the GCD circuit.
You will need two registers to hold A and B, a pair of subtractors
(you can use the Logisim arithmetic library) and a few multiplexers and
some gates for the registers’ enable networks. You can also use a library
comparator to avoid having to subtract and analyse the sign, etc.
Please use the design board provided in 7COM1013.circ-T04L1e07.circ
attached to the email. The input pins A and B are required for setting
the initial values of A and B respectively. The values can be asserted
by the tester in any clock cycle, in which case the tester also raises the
input pin Set in the same clock cycle. The output GCD is examined by
the tester when your solution circuit raises the output pin done.
Please use the clock signal provided via the corresponding pin.